+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Form Label Component
 - Form RadioButton Component
 - Checkboxes Component
+- FormMultiTextField Component
 
 ### Changed